Updates for the IAU during cycling, along with two new python scripts under tools to compute analysis increments in netcdf.

+++ forecast.csh
@@ -1,5 +1,5 @@
 #!/bin/csh -f
-
+# forecast.csh
 date
 
 # Process arguments
@@ -50,6 +50,11 @@ set config_run_duration = 0_${self_fcLengthHR}:00:00
 set output_interval = 0_${self_fcIntervalHR}:00:00
 set deleteZerothForecast = deleteZerothForecastTEMPLATE
 
+# IAU
+set IAU_on = False 		# True or False (=default)
+set StartDate = ${NMLDate}
+set self_CycleFreqHR = fcIntervalHRTEMPLATE
+
 # static variables
 set self_icStatePrefix = ${ANFilePrefix}
 set self_ModelConfigDir = $forecastModelConfigDir
@@ -60,14 +65,15 @@ set self_ModelConfigDir = $forecastModelConfigDir
 rm ${localStaticFieldsPrefix}*.nc
 rm ${localStaticFieldsPrefix}*.nc-lock
 set localStaticFieldsFile = ${localStaticFieldsFileOuter}
-rm ${localStaticFieldsFile}
+if( -e ${localStaticFieldsFile} ) rm ${localStaticFieldsFile}
 set icFileExt = ${fileDate}.nc
 set icFile = ${ICFilePrefix}.${icFileExt}
-rm ./${icFile}
+if( -e ${icFile} ) rm ./${icFile}
 if ( ${InitializationType} == "ColdStart" && ${thisValidDate} == ${FirstCycleDate}) then
   set initialState = ${InitICDir}/${InitFilePrefixOuter}.${icFileExt}
   set do_DAcycling = "false"
   ln -sfv ${initialState} ${localStaticFieldsFile}
+  setenv IAU False
 else
   set initialState = ${self_icStateDir}/${self_icStatePrefix}.${icFileExt}
   set do_DAcycling = "true"
@@ -75,8 +81,39 @@ else
   set memberStaticFieldsFile = ${StaticFieldsDirOuter}${StaticMemDir}/${StaticFieldsFileOuter}
   ln -sfv ${memberStaticFieldsFile} ${localStaticFieldsFile}${OrigFileSuffix}
   cp -v ${memberStaticFieldsFile} ${localStaticFieldsFile}
+
+  if ( ${IAU} == True ) then
+   set BGFileExt = `$TimeFmtChange ${IAUDate}`.00.00.nc   	# analysis - 3h [YYYY-MM-DD_HH.00.00]
+   set BGFile   = ${prevCyclingFCDir}/${FCFilePrefix}.${BGFileExt}	# mpasout at (analysis - 3h)
+   set BGFileA  = ${CyclingDAInDir}/${BGFilePrefix}.${icFileExt}	# bg at the analysis time
+   echo ""
+   echo "IAU needs two background files:"
+   echo "IC: ${BGFile}"
+   echo "bg: ${BGFileA}"
+
+   if ( -e ${BGFile} && -e ${BGFileA} ) then
+    echo "IAU starts from ${IAUDate}."
+    set IAU_on = True
+    set self_CycleFreqHR = CyclingFrequencyHR
+    set StartDate  = `$TimeFmtChange ${IAUDate}`:00:00	# YYYYMMDDHH => YYYY-MM-DD_HH:00:00
+    # Compute analysis increments (AmB)
+    ln -sfv ${initialState} ${ANFilePrefix}.${icFileExt}		# an.YYYY-MM-DD_HH.00.00.nc
+    ln -sfv ${BGFileA}      ${BGFilePrefix}.${icFileExt}		# bg.YYYY-MM-DD_HH.00.00.nc
+    setenv myCommand "${create_amb_in_nc} ${thisValidDate}" # ${IAU_window_s}"
+    echo "$myCommand"
+    ${myCommand}
+    set famb = AmB.`$TimeFmtChange ${IAUDate}`.00.00.nc
+    ls -lL $famb						|| exit
+    # Initial condition (mpasin.YYYY-MM-DD_HH.00.00.nc)
+    ln -sfv ${BGFile} ${ICFilePrefix}.${BGFileExt}		|| exit
+   else		# either analysis or background does not exist; IAU is off.
+    echo "IAU was on, but no input files. So it is off and initialized at ${fileDate}."
+    setenv IAU False
+   endif
+  endif # ( ${IAU} == True ) then
 endif
 ln -sfv ${initialState} ./${icFile}
+if ( ${IAU} == True ) \rm -f ./${icFile}
 
 ## link MPAS mesh graph info
 rm ./x1.${MPASnCellsOuter}.graph.info*
@@ -93,12 +130,12 @@ foreach staticfile ( \
 stream_list.${MPASCore}.surface \
 stream_list.${MPASCore}.diagnostics \
 )
-  rm ./$staticfile
+  if( -e $staticfile ) rm ./$staticfile
   ln -sfv $self_ModelConfigDir/$staticfile .
 end
 
 ## copy/modify dynamic streams file
-rm ${StreamsFile}
+if( -e ${StreamsFile}) rm ${StreamsFile}
 cp -v $self_ModelConfigDir/${StreamsFile} .
 sed -i 's@nCells@'${MPASnCellsOuter}'@' ${StreamsFile}
 sed -i 's@outputInterval@'${output_interval}'@' ${StreamsFile}
@@ -108,14 +145,21 @@ sed -i 's@FCFilePrefix@'${FCFilePrefix}'@' ${StreamsFile}
 sed -i 's@forecastPrecision@'${forecastPrecision}'@' ${StreamsFile}
 
 ## copy/modify dynamic namelist
-rm ${NamelistFile}
+if( -e ${NamelistFile}) rm ${NamelistFile}
 cp -v ${self_ModelConfigDir}/${NamelistFile} .
-sed -i 's@startTime@'${NMLDate}'@' $NamelistFile
+sed -i 's@startTime@'${StartDate}'@' $NamelistFile
 sed -i 's@fcLength@'${config_run_duration}'@' $NamelistFile
 sed -i 's@nCells@'${MPASnCellsOuter}'@' $NamelistFile
 sed -i 's@modelDT@'${MPASTimeStep}'@' $NamelistFile
 sed -i 's@diffusionLengthScale@'${MPASDiffusionLengthScale}'@' $NamelistFile
 sed -i 's@configDODACycling@'${do_DAcycling}'@' $NamelistFile
+if ( ${IAU_on} == True ) then
+    sed -i 's@off@on@' $NamelistFile
+    echo "IAU is turned on."
+endif
+#   set IAU_window_s = 3600.
+#   @ IAU_window_s *= ${self_fcIntervalHR}
+#   sed -i 's@IAUwindowSecs@'${IAU_window_s}'@' $NamelistFile
 
 if ( ${self_fcLengthHR} == 0 ) then
   ## zero-length forecast case (NOT CURRENTLY USED)
@@ -130,15 +174,10 @@ else
   set fcDate = `$advanceCYMDH ${thisValidDate} ${self_fcIntervalHR}`
   set finalFCDate = `$advanceCYMDH ${thisValidDate} ${self_fcLengthHR}`
   while ( ${fcDate} <= ${finalFCDate} )
-    set yy = `echo ${fcDate} | cut -c 1-4`
-    set mm = `echo ${fcDate} | cut -c 5-6`
-    set dd = `echo ${fcDate} | cut -c 7-8`
-    set hh = `echo ${fcDate} | cut -c 9-10`
-    set fcFileDate  = ${yy}-${mm}-${dd}_${hh}.00.00
-    set fcFileExt = ${fcFileDate}.nc
-    set fcFile = ${FCFilePrefix}.${fcFileExt}
+    set fcFileDate  = `$TimeFmtChange ${fcDate}`
+    set fcFile = ${FCFilePrefix}.${fcFileDate}.00.00.nc
 
-    rm ${fcFile}
+    if( -e ${fcFile} ) rm ${fcFile}
 
     set fcDate = `$advanceCYMDH ${fcDate} ${self_fcIntervalHR}`
     setenv fcDate ${fcDate}
@@ -146,7 +185,7 @@ else
 
   # Run the executable
   # ==================
-  rm ./${ForecastEXE}
+  if( -e ${ForecastEXE} ) rm ./${ForecastEXE}
   ln -sfv ${ForecastBuildDir}/${ForecastEXE} ./
   mpiexec ./${ForecastEXE}
 
@@ -163,22 +202,17 @@ else
   if ( ${InitializationType} == "WarmStart" ) then
     rm ${localStaticFieldsFile}
     mv ${localStaticFieldsFile}${OrigFileSuffix} ${localStaticFieldsFile}
-  endid
+  endif
 endif
 
 if ( "$deleteZerothForecast" == "True" ) then
   # Optionally remove initial forecast file
   # =======================================
   set fcDate = ${thisValidDate}
-  set yy = `echo ${fcDate} | cut -c 1-4`
-  set mm = `echo ${fcDate} | cut -c 5-6`
-  set dd = `echo ${fcDate} | cut -c 7-8`
-  set hh = `echo ${fcDate} | cut -c 9-10`
-  set fcFileDate  = ${yy}-${mm}-${dd}_${hh}.00.00
-  set fcFileExt = ${fcFileDate}.nc
-  set fcFile = ${FCFilePrefix}.${fcFileExt}
+  set fcFileDate  = `$TimeFmtChange ${fcDate}`
+  set fcFile = ${FCFilePrefix}.${fcFileDate}.00.00.nc
   rm ${fcFile}
-  set diagFile = ${DIAGFilePrefix}.${fcFileExt}
+  set diagFile = ${DIAGFilePrefix}.${fcFileDate}.00.00.nc
   rm ${diagFile}
 endif
 
@@ -187,14 +221,15 @@ endif
 #TODO: do this in a separate post-processing script
 #      either in parallel or using only single processor
 #      instead of full set of job processors
+if ( ${IAU} == True ) then
+set fcDate = `$advanceCYMDH ${IAUDate} ${self_fcIntervalHR}`
+set finalFCDate = `$advanceCYMDH ${IAUDate} ${self_fcLengthHR}`
+else
 set fcDate = `$advanceCYMDH ${thisValidDate} ${self_fcIntervalHR}`
 set finalFCDate = `$advanceCYMDH ${thisValidDate} ${self_fcLengthHR}`
+endif
 while ( ${fcDate} <= ${finalFCDate} )
-  set yy = `echo ${fcDate} | cut -c 1-4`
-  set mm = `echo ${fcDate} | cut -c 5-6`
-  set dd = `echo ${fcDate} | cut -c 7-8`
-  set hh = `echo ${fcDate} | cut -c 9-10`
-  set fcFileDate  = ${yy}-${mm}-${dd}_${hh}.00.00
+  set fcFileDate  = `$TimeFmtChange ${fcDate}`.00.00
   set fcFileExt = ${fcFileDate}.nc
   set fcFile = ${FCFilePrefix}.${fcFileExt}
 
@@ -203,6 +238,7 @@ while ( ${fcDate} <= ${finalFCDate} )
     # first try member-specific state file (central GFS state when ArgMember==0)
     set seaMemDir = `${memberDir} ens $ArgMember "${seaMemFmt}" -m ${seaMaxMembers}`
     set SeaFile = ${SeaAnaDir}/${fcDate}${seaMemDir}/${SeaFilePrefix}.${fcFileExt}
+    if ( -e ${SeaFile} ) then
     ncks -A -v ${MPASSeaVariables} ${SeaFile} ${fcFile}
     if ( $status != 0 ) then
       echo "WARNING in $0 : ncks -A -v ${MPASSeaVariables} ${SeaFile} ${fcFile}" > ./WARNING
@@ -217,6 +253,9 @@ while ( ${fcDate} <= ${finalFCDate} )
         exit 1
       endif
     endif
+    else
+      echo "WARNING in $0 : ${SeaFile} is not found. Skip adding ${MPASSeaVariables} to $fcFile" > ./WARNING
+    endif # ( -e $${SeaFile} ) then
   endif
 
   ## Add MPASJEDIDiagVariables to the next cycle bg file (if needed)
@@ -229,6 +268,7 @@ while ( ${fcDate} <= ${finalFCDate} )
   end
   set diagFile = ${DIAGFilePrefix}.${fcFileExt}
   if ( $copyDiags > 0 ) then
+    echo "ncks -A -v ${MPASJEDIDiagVariables} ${diagFile} ${fcFile}"
     ncks -A -v ${MPASJEDIDiagVariables} ${diagFile} ${fcFile}
   endif
   # rm ${diagFile}
